Added my own fork with fixes for tar symlinks
Fixed arc format bug with switch case fallthrough

+++ packages/archiver_ffi/native/archiver_lib/pack.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"github.com/mholt/archiver"
+	"github.com/ganeshrvel/archiver"
 	"log"
 	"os"
 	"path/filepath"
@@ -29,6 +29,7 @@ func (arc CommonArchive) doPack() error {
 
 	_filename := arc.meta.filename
 	_fileList := arc.pack.fileList
+	_overwriteExisting := arc.pack.overwriteExisting
 
 	arcFileObj, err := archiver.ByExtension(_filename)
 
@@ -36,7 +37,7 @@ func (arc CommonArchive) doPack() error {
 		return err
 	}
 
-	err = archiveFormat(&arcFileObj, "")
+	err = archiveFormat(&arcFileObj, "", _overwriteExisting)
 
 	if err != nil {
 		return err
@@ -74,11 +75,13 @@ func startPacking(meta *ArchiveMeta, pack *ArchivePack) error {
 	_meta := *meta
 	_pack := *pack
 
+	_overwriteExisting := pack.overwriteExisting
+
 	var arcPackObj ArchivePacker
 
 	ext := filepath.Ext(_meta.filename)
 
-	if fileExists(_meta.filename) {
+	if _overwriteExisting && fileExists(_meta.filename) {
 		if err := os.Remove(_meta.filename); err != nil {
 			return err
 		}
